Changes to FEApp aztecoo epetra phalanx trilinoscouplings xpetra so that almost all tests possibly affected by Epetra 64 bit changes pass.  Most of the issues are related to the ambiguity created by int/long long overloading and the passed in type something else (like unsigned int).  Other issue fixed here is to use int based maps unless there is a reason to use long long map.

+++ packages/trilinoscouplings/examples/scaling/example_Darcy.cpp
@@ -1697,7 +1697,7 @@ int main(int argc, char *argv[]) {
     // ********  Calculate Error in Solution ***************
    // Import solution onto current processor
 
-     Epetra_Map  solnMap(numFacesGlobal+numNodesGlobal, numFacesGlobal+numNodesGlobal, 0, Comm);
+     Epetra_Map  solnMap(static_cast<int>(numFacesGlobal+numNodesGlobal), static_cast<int>(numFacesGlobal+numNodesGlobal), 0, Comm);
      Epetra_Import  solnImporter(solnMap, globalMapJoint);
      Epetra_FEVector  uCoeff(solnMap);
      uCoeff.Import(globalSoln, solnImporter, Insert);
